requirements.txt
click
g4f
GitPython
